chore: move leasing packages into a subfolder
