+++ src/core/index.ts
@@ -31,7 +31,7 @@ import { Task } from './task.ts'
 import * as helpers from './helpers.ts'
 import { DependencyManager } from './dependency_managers/index.ts'
 import { AccountManager } from './account_manager.ts'
-import { LeaseManager } from './lease_manager.ts'
+import { LeaseManager } from './lease/lease_manager.ts'
 import { CertificateManager } from './certificate_manager.ts'
 
 // Expose components from the core module
