+++ src/core/lease/lease_manager.ts
@@ -14,11 +14,11 @@
  * limitations under the License.
  *
  */
-import { MissingArgumentError } from './errors.ts'
-import { flags } from '../commands/index.ts'
-import type { ConfigManager } from './config_manager.ts'
-import type { K8 } from './k8.ts'
-import type { SoloLogger } from './logging.ts'
+import { MissingArgumentError } from '../errors.ts'
+import { flags } from '../../commands/index.ts'
+import type { ConfigManager } from '../config_manager.ts'
+import type { K8 } from '../k8.ts'
+import type { SoloLogger } from '../logging.ts'
 import { type LeaseRenewalService } from './lease_renewal.ts'
 import { Lease } from './lease.ts'
 import { LeaseHolder } from './lease_holder.ts'
